storage: move some structs into sink::kafka::util
